+++ src/game/hustles.js
@@ -1,10 +1,10 @@
-import { createInstantHustle } from './content/schema.js';
 import { getInstantHustleDefinitions } from './hustles/definitions/instantHustles.js';
 import { createKnowledgeHustles } from './hustles/knowledgeHustles.js';
+import { ACTIONS, INSTANT_ACTIONS, STUDY_ACTIONS } from './actions/definitions.js';
 
-const INSTANT_HUSTLES = getInstantHustleDefinitions().map(definition => createInstantHustle(definition));
-const KNOWLEDGE_HUSTLES = createKnowledgeHustles();
-
-export const HUSTLES = [...INSTANT_HUSTLES, ...KNOWLEDGE_HUSTLES];
+const MARKET_HUSTLES = getInstantHustleDefinitions();
+export const HUSTLES = MARKET_HUSTLES;
+export const KNOWLEDGE_HUSTLES = createKnowledgeHustles();
+export { ACTIONS, INSTANT_ACTIONS, STUDY_ACTIONS };
 
 export * from './hustles/helpers.js';